issue-92: added OrganizationalStructureController
