+++ examples/EmployeesMvc/webapp/Views/Shared/_Layout.cshtml
@@ -39,12 +39,12 @@
                                     @_appSettings.Translate(languageName, "OrganizationalStructure")
                                 </a>
                                 <div class="dropdown-menu" aria-labelledby="navbarDropdown">
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="BriefDescription">@_appSettings.Translate(languageName, "BriefDescription")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="OrganizationalStructure" asp-action="BriefDescription">@_appSettings.Translate(languageName, "BriefDescription")</a>
                                     <div class="dropdown-divider"></div>
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="Organizations">@_appSettings.Translate(languageName, "Organizations")</a>
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="Departments">@_appSettings.Translate(languageName, "Departments")</a>
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="Teams">@_appSettings.Translate(languageName, "Teams")</a>
-                                    <a class="dropdown-item" asp-area="" asp-controller="Home" asp-action="JobPositions">@_appSettings.Translate(languageName, "JobPositions")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="OrganizationalStructure" asp-action="Organizations">@_appSettings.Translate(languageName, "Organizations")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="OrganizationalStructure" asp-action="Departments">@_appSettings.Translate(languageName, "Departments")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="OrganizationalStructure" asp-action="Teams">@_appSettings.Translate(languageName, "Teams")</a>
+                                    <a class="dropdown-item" asp-area="" asp-controller="OrganizationalStructure" asp-action="JobPositions">@_appSettings.Translate(languageName, "JobPositions")</a>
                                 </div>
                             </li>
                             <li class="nav-item dropdown">
